abstract common logic for reuse
