+++ core/src/main/java/com/redhat/lightblue/client/response/DefaultLightblueDataResponse.java
@@ -2,112 +2,31 @@ package com.redhat.lightblue.client.response;
 
 import java.io.IOException;
 import java.lang.reflect.Array;
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.fasterxml.jackson.databind.node.NullNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.redhat.lightblue.client.model.DataError;
-import com.redhat.lightblue.client.model.Error;
-import com.redhat.lightblue.client.util.JSON;
 
-public class DefaultLightblueDataResponse extends AbstractLightblueResponse implements LightblueDataResponse, LightblueErrorResponse {
+public class DefaultLightblueDataResponse extends DefaultLightblueErrorResponse implements LightblueDataResponse {
 
-    public DefaultLightblueDataResponse(JsonNode responseNode) throws LightblueResponseException {
-        this(responseNode, JSON.getDefaultObjectMapper());
+    public DefaultLightblueDataResponse(JsonNode responseNode) 
+            throws LightblueResponseException {
+        super(responseNode);
     }
 
-    public DefaultLightblueDataResponse(JsonNode responseNode, ObjectMapper mapper) throws LightblueResponseException {
+    public DefaultLightblueDataResponse(JsonNode responseNode, ObjectMapper mapper) 
+            throws LightblueResponseException {
         super(responseNode, mapper);
-
-        if ((getJson() == null) || hasAnyErrors()) {
-            throw new LightblueResponseException("Lightblue exception occurred: ", this);
-        }
     }
 
-    public DefaultLightblueDataResponse(String responseText) throws LightblueParseException, LightblueResponseException {
-        this(responseText, JSON.getDefaultObjectMapper());
+    public DefaultLightblueDataResponse(String responseText) 
+            throws LightblueParseException, LightblueResponseException {
+        super(responseText);
     }
 
-    public DefaultLightblueDataResponse(String responseText, ObjectMapper mapper) throws LightblueParseException, LightblueResponseException {
+    public DefaultLightblueDataResponse(String responseText, ObjectMapper mapper) 
+            throws LightblueParseException, LightblueResponseException {
         super(responseText, mapper);
-
-        if (hasAnyErrors()) {
-            throw new LightblueResponseException("Lightblue exception occurred: ", this);
-        }
-    }
-
-    @Override
-    public boolean hasDataErrors() {
-        if (getJson() == null) {
-            return false;
-        }
-
-        JsonNode err = getJson().get("dataErrors");
-        return err != null && !(err instanceof NullNode) && err.size() > 0;
-    }
-
-    @Override
-    public boolean hasLightblueErrors() {
-        if (getJson() == null) {
-            return false;
-        }
-
-        JsonNode objectTypeNode = getJson().get("status");
-        if(objectTypeNode != null
-                && (objectTypeNode.textValue().equalsIgnoreCase("error")
-                  || objectTypeNode.textValue().equalsIgnoreCase("partial"))){
-            return true;
-        }
-
-        JsonNode err = getJson().get("errors");
-        return err != null && !(err instanceof NullNode) && err.size() > 0;
-    }
-
-    public boolean hasAnyErrors() {
-        return hasDataErrors() || hasLightblueErrors();
-    }
-
-    @Override
-    public DataError[] getDataErrors() {
-        List<DataError> list = new ArrayList<>();
-        if (getJson() == null) {
-            return null;
-        }
-        JsonNode err = getJson().get("dataErrors");
-        if (err instanceof ObjectNode) {
-            list.add(DataError.fromJson((ObjectNode) err));
-        } else if (err instanceof ArrayNode) {
-            for (Iterator<JsonNode> itr = ((ArrayNode) err).elements(); itr.hasNext();) {
-                list.add(DataError.fromJson((ObjectNode) itr.next()));
-            }
-        } else {
-            return null;
-        }
-        return list.toArray(new DataError[list.size()]);
-    }
-
-    @Override
-    public Error[] getLightblueErrors() {
-        List<Error> list = new ArrayList<>();
-        if (getJson() == null) {
-            return null;
-        }
-        JsonNode err = getJson().get("errors");
-        if (err instanceof ObjectNode) {
-            list.add(Error.fromJson(err));
-        } else if (err instanceof ArrayNode) {
-            for (Iterator<JsonNode> itr = ((ArrayNode) err).elements(); itr.hasNext();) {
-                list.add(Error.fromJson(itr.next()));
-            }
-        } else {
-            return null;
-        }
-        return list.toArray(new Error[list.size()]);
     }
 
     @Override
